fix: 음악 선택 context -> zustand
